refactor proof system
