+++ zokrates_core/src/proof_system/libsnark/gm17.rs
@@ -1,34 +1,13 @@
-use ir;
+use ir::{Prog, Witness};
 use proof_system::libsnark::ffi::{Buffer, ProofResult, SetupResult};
-use proof_system::libsnark::{prepare_generate_proof, prepare_public_inputs, prepare_setup};
-use proof_system::solidity::{
-    SolidityAbi, SOLIDITY_G2_ADDITION_LIB, SOLIDITY_PAIRING_LIB, SOLIDITY_PAIRING_LIB_V2,
+use proof_system::libsnark::{
+    prepare_generate_proof, prepare_public_inputs, prepare_setup, Libsnark,
 };
-use proof_system::{G1Affine, G2Affine, Proof, ProofSystem, SetupKeypair};
-use regex::Regex;
-
+use proof_system::scheme::gm17::GM17;
+use proof_system::scheme::Scheme;
+use proof_system::{Backend, Proof, SetupKeypair};
 use zokrates_field::Bn128Field;
 use zokrates_field::Field;
-
-pub struct GM17 {}
-
-#[derive(Serialize, Deserialize)]
-pub struct VerificationKey {
-    h: G2Affine,
-    g_alpha: G1Affine,
-    h_beta: G2Affine,
-    g_gamma: G1Affine,
-    h_gamma: G2Affine,
-    query: Vec<G1Affine>,
-    raw: String,
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct ProofPoints {
-    a: G1Affine,
-    b: G2Affine,
-    c: G1Affine,
-}
 
 extern "C" {
     fn gm17_bn128_setup(
@@ -59,11 +38,10 @@ extern "C" {
     ) -> bool;
 }
 
-impl ProofSystem<Bn128Field> for GM17 {
-    type VerificationKey = VerificationKey;
-    type ProofPoints = ProofPoints;
-
-    fn setup(program: ir::Prog<Bn128Field>) -> SetupKeypair<VerificationKey> {
+impl Backend<Bn128Field, GM17> for Libsnark {
+    fn setup(
+        program: Prog<Bn128Field>,
+    ) -> SetupKeypair<<GM17 as Scheme<Bn128Field>>::VerificationKey> {
         let (a_arr, b_arr, c_arr, a_vec, b_vec, c_vec, num_constraints, num_variables, num_inputs) =
             prepare_setup(program);
 
@@ -97,10 +75,10 @@ impl ProofSystem<Bn128Field> for GM17 {
     }
 
     fn generate_proof(
-        program: ir::Prog<Bn128Field>,
-        witness: ir::Witness<Bn128Field>,
+        program: Prog<Bn128Field>,
+        witness: Witness<Bn128Field>,
         proving_key: Vec<u8>,
-    ) -> Proof<ProofPoints> {
+    ) -> Proof<<GM17 as Scheme<Bn128Field>>::ProofPoints> {
         let (public_inputs_arr, public_inputs_length, private_inputs_arr, private_inputs_length) =
             prepare_generate_proof(program, witness);
 
@@ -130,114 +108,12 @@ impl ProofSystem<Bn128Field> for GM17 {
         serde_json::from_str(String::from_utf8(proof).unwrap().as_str()).unwrap()
     }
 
-    fn export_solidity_verifier(vk: VerificationKey, abi: SolidityAbi) -> String {
-        let (mut template_text, solidity_pairing_lib) = match abi {
-            SolidityAbi::V1 => (
-                String::from(CONTRACT_TEMPLATE),
-                String::from(SOLIDITY_PAIRING_LIB),
-            ),
-            SolidityAbi::V2 => (
-                String::from(CONTRACT_TEMPLATE_V2),
-                String::from(SOLIDITY_PAIRING_LIB_V2),
-            ),
-        };
-
-        // replace things in template
-        let vk_regex = Regex::new(r#"(<%vk_[^i%]*%>)"#).unwrap();
-        let vk_query_len_regex = Regex::new(r#"(<%vk_query_length%>)"#).unwrap();
-        let vk_query_repeat_regex = Regex::new(r#"(<%vk_query_pts%>)"#).unwrap();
-        let vk_input_len_regex = Regex::new(r#"(<%vk_input_length%>)"#).unwrap();
-        let input_loop = Regex::new(r#"(<%input_loop%>)"#).unwrap();
-        let input_argument = Regex::new(r#"(<%input_argument%>)"#).unwrap();
-
-        template_text = vk_regex
-            .replace(template_text.as_str(), vk.h.to_string().as_str())
-            .into_owned();
-
-        template_text = vk_regex
-            .replace(template_text.as_str(), vk.g_alpha.to_string().as_str())
-            .into_owned();
-
-        template_text = vk_regex
-            .replace(template_text.as_str(), vk.h_beta.to_string().as_str())
-            .into_owned();
-
-        template_text = vk_regex
-            .replace(template_text.as_str(), vk.g_gamma.to_string().as_str())
-            .into_owned();
-
-        template_text = vk_regex
-            .replace(template_text.as_str(), vk.h_gamma.to_string().as_str())
-            .into_owned();
-
-        let query_count: usize = vk.query.len();
-        template_text = vk_query_len_regex
-            .replace(template_text.as_str(), format!("{}", query_count).as_str())
-            .into_owned();
-
-        template_text = vk_input_len_regex
-            .replace(
-                template_text.as_str(),
-                format!("{}", query_count - 1).as_str(),
-            )
-            .into_owned();
-
-        // feed input values only if there are any
-        template_text = if query_count > 1 {
-            input_loop.replace(
-                template_text.as_str(),
-                r#"
-        for(uint i = 0; i < input.length; i++){
-            inputValues[i] = input[i];
-        }"#,
-            )
-        } else {
-            input_loop.replace(template_text.as_str(), "")
-        }
-        .to_string();
-
-        // take input values as argument only if there are any
-        template_text = if query_count > 1 {
-            input_argument.replace(
-                template_text.as_str(),
-                format!(", uint[{}] memory input", query_count - 1).as_str(),
-            )
-        } else {
-            input_argument.replace(template_text.as_str(), "")
-        }
-        .to_string();
-
-        let mut query_repeat_text = String::new();
-        for (i, g1) in vk.query.iter().enumerate() {
-            query_repeat_text.push_str(
-                format!(
-                    "vk.query[{}] = Pairing.G1Point({});",
-                    i,
-                    g1.to_string().as_str()
-                )
-                .as_str(),
-            );
-            if i < query_count - 1 {
-                query_repeat_text.push_str("\n        ");
-            }
-        }
-
-        template_text = vk_query_repeat_regex
-            .replace(template_text.as_str(), query_repeat_text.as_str())
-            .into_owned();
-
-        let re = Regex::new(r"(?P<v>0[xX][0-9a-fA-F]{64})").unwrap();
-        template_text = re.replace_all(&template_text, "uint256($v)").to_string();
-
-        format!(
-            "{}{}{}",
-            SOLIDITY_G2_ADDITION_LIB, solidity_pairing_lib, template_text
-        )
-    }
-
-    fn verify(vk: VerificationKey, proof: Proof<ProofPoints>) -> bool {
-        let vk_raw = hex::decode(vk.raw.clone()).unwrap();
-        let proof_raw = hex::decode(proof.raw.unwrap()).unwrap();
+    fn verify(
+        vk: <GM17 as Scheme<Bn128Field>>::VerificationKey,
+        proof: Proof<<GM17 as Scheme<Bn128Field>>::ProofPoints>,
+    ) -> bool {
+        let vk_raw = hex::decode(vk.raw.unwrap().clone()).unwrap();
+        let proof_raw = hex::decode(proof.raw.unwrap().clone()).unwrap();
 
         let public_inputs: Vec<_> = proof
             .inputs
@@ -265,133 +141,3 @@ impl ProofSystem<Bn128Field> for GM17 {
         }
     }
 }
-
-const CONTRACT_TEMPLATE_V2: &str = r#"
-contract Verifier {
-    using Pairing for *;
-    struct VerifyingKey {
-        Pairing.G2Point h;
-        Pairing.G1Point g_alpha;
-        Pairing.G2Point h_beta;
-        Pairing.G1Point g_gamma;
-        Pairing.G2Point h_gamma;
-        Pairing.G1Point[] query;
-    }
-    struct Proof {
-        Pairing.G1Point a;
-        Pairing.G2Point b;
-        Pairing.G1Point c;
-    }
-    function verifyingKey() pure internal returns (VerifyingKey memory vk) {
-        vk.h= Pairing.G2Point(<%vk_h%>);
-        vk.g_alpha = Pairing.G1Point(<%vk_g_alpha%>);
-        vk.h_beta = Pairing.G2Point(<%vk_h_beta%>);
-        vk.g_gamma = Pairing.G1Point(<%vk_g_gamma%>);
-        vk.h_gamma = Pairing.G2Point(<%vk_h_gamma%>);
-        vk.query = new Pairing.G1Point[](<%vk_query_length%>);
-        <%vk_query_pts%>
-    }
-    function verify(uint[] memory input, Proof memory proof) internal view returns (uint) {
-        uint256 snark_scalar_field = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
-        VerifyingKey memory vk = verifyingKey();
-        require(input.length + 1 == vk.query.length);
-        // Compute the linear combination vk_x
-        Pairing.G1Point memory vk_x = Pairing.G1Point(0, 0);
-        for (uint i = 0; i < input.length; i++) {
-            require(input[i] < snark_scalar_field);
-            vk_x = Pairing.addition(vk_x, Pairing.scalar_mul(vk.query[i + 1], input[i]));
-        }
-        vk_x = Pairing.addition(vk_x, vk.query[0]);
-        /**
-         * e(A*G^{alpha}, B*H^{beta}) = e(G^{alpha}, H^{beta}) * e(G^{psi}, H^{gamma})
-         *                              * e(C, H)
-         * where psi = \sum_{i=0}^l input_i pvk.query[i]
-         */
-        if (!Pairing.pairingProd4(vk.g_alpha, vk.h_beta, vk_x, vk.h_gamma, proof.c, vk.h, Pairing.negate(Pairing.addition(proof.a, vk.g_alpha)), Pairing.addition(proof.b, vk.h_beta))) return 1;
-        /**
-         * e(A, H^{gamma}) = e(G^{gamma}, B)
-         */
-        if (!Pairing.pairingProd2(proof.a, vk.h_gamma, Pairing.negate(vk.g_gamma), proof.b)) return 2;
-        return 0;
-    }
-    function verifyTx(
-            Proof memory proof<%input_argument%>
-        ) public view returns (bool r) {
-        uint[] memory inputValues = new uint[](input.length);
-        <%input_loop%>
-        if (verify(inputValues, proof) == 0) {
-            return true;
-        } else {
-            return false;
-        }
-    }
-}
-"#;
-
-const CONTRACT_TEMPLATE: &str = r#"
-contract Verifier {
-    using Pairing for *;
-    struct VerifyingKey {
-        Pairing.G2Point h;
-        Pairing.G1Point g_alpha;
-        Pairing.G2Point h_beta;
-        Pairing.G1Point g_gamma;
-        Pairing.G2Point h_gamma;
-        Pairing.G1Point[] query;
-    }
-    struct Proof {
-        Pairing.G1Point a;
-        Pairing.G2Point b;
-        Pairing.G1Point c;
-    }
-    function verifyingKey() pure internal returns (VerifyingKey memory vk) {
-        vk.h = Pairing.G2Point(<%vk_h%>);
-        vk.g_alpha = Pairing.G1Point(<%vk_g_alpha%>);
-        vk.h_beta = Pairing.G2Point(<%vk_h_beta%>);
-        vk.g_gamma = Pairing.G1Point(<%vk_g_gamma%>);
-        vk.h_gamma = Pairing.G2Point(<%vk_h_gamma%>);
-        vk.query = new Pairing.G1Point[](<%vk_query_length%>);
-        <%vk_query_pts%>
-    }
-    function verify(uint[] memory input, Proof memory proof) internal view returns (uint) {
-        uint256 snark_scalar_field = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
-        VerifyingKey memory vk = verifyingKey();
-        require(input.length + 1 == vk.query.length);
-        // Compute the linear combination vk_x
-        Pairing.G1Point memory vk_x = Pairing.G1Point(0, 0);
-        for (uint i = 0; i < input.length; i++) {
-            require(input[i] < snark_scalar_field);
-            vk_x = Pairing.addition(vk_x, Pairing.scalar_mul(vk.query[i + 1], input[i]));
-        }
-        vk_x = Pairing.addition(vk_x, vk.query[0]);
-        /**
-         * e(A*G^{alpha}, B*H^{beta}) = e(G^{alpha}, H^{beta}) * e(G^{psi}, H^{gamma})
-         *                              * e(C, H)
-         * where psi = \sum_{i=0}^l input_i pvk.query[i]
-         */
-        if (!Pairing.pairingProd4(vk.g_alpha, vk.h_beta, vk_x, vk.h_gamma, proof.c, vk.h, Pairing.negate(Pairing.addition(proof.a, vk.g_alpha)), Pairing.addition(proof.b, vk.h_beta))) return 1;
-        /**
-         * e(A, H^{gamma}) = e(G^{gamma}, b)
-         */
-        if (!Pairing.pairingProd2(proof.a, vk.h_gamma, Pairing.negate(vk.g_gamma), proof.b)) return 2;
-        return 0;
-    }
-    function verifyTx(
-            uint[2] memory a,
-            uint[2][2] memory b,
-            uint[2] memory c<%input_argument%>
-        ) public view returns (bool r) {
-        Proof memory proof;
-        proof.a = Pairing.G1Point(a[0], a[1]);
-        proof.b = Pairing.G2Point([b[0][0], b[0][1]], [b[1][0], b[1][1]]);
-        proof.c = Pairing.G1Point(c[0], c[1]);
-        uint[] memory inputValues = new uint[](input.length);
-        <%input_loop%>
-        if (verify(inputValues, proof) == 0) {
-            return true;
-        } else {
-            return false;
-        }
-    }
-}
-"#;
