+++ zokrates_core/src/proof_system/libsnark/mod.rs
@@ -8,6 +8,8 @@ use std::cmp::max;
 use std::collections::HashMap;
 use zokrates_field::Field;
 
+pub struct Libsnark;
+
 // utility function. Converts a Field's vector-based byte representation to fixed size array.
 fn vec_as_u8_32_array(vec: &Vec<u8>) -> [u8; 32] {
     assert!(vec.len() <= 32);
