+++ zokrates_core/src/proof_system/scheme/groth16.rs
@@ -1,0 +1,254 @@
+use proof_system::scheme::{Scheme, SolidityCompatibleScheme};
+use proof_system::solidity::{
+    SolidityAbi, SOLIDITY_G2_ADDITION_LIB, SOLIDITY_PAIRING_LIB, SOLIDITY_PAIRING_LIB_V2,
+};
+use proof_system::{G1Affine, G2Affine};
+use regex::Regex;
+use zokrates_field::Field;
+
+pub struct G16;
+
+#[derive(Serialize, Deserialize)]
+pub struct ProofPoints<G1, G2> {
+    pub a: G1,
+    pub b: G2,
+    pub c: G1,
+}
+
+#[derive(Serialize, Deserialize)]
+pub struct VerificationKey<G1, G2> {
+    pub alpha: G1,
+    pub beta: G2,
+    pub gamma: G2,
+    pub delta: G2,
+    pub gamma_abc: Vec<G1>,
+}
+
+impl<T: Field> Scheme<T> for G16 {
+    type VerificationKey = VerificationKey<G1Affine, G2Affine>;
+    type ProofPoints = ProofPoints<G1Affine, G2Affine>;
+}
+
+impl<T: Field> SolidityCompatibleScheme<T> for G16 {
+    fn export_solidity_verifier(
+        vk: <G16 as Scheme<T>>::VerificationKey,
+        abi: SolidityAbi,
+    ) -> String {
+        let (mut template_text, solidity_pairing_lib) = match abi {
+            SolidityAbi::V1 => (
+                String::from(CONTRACT_TEMPLATE),
+                String::from(SOLIDITY_PAIRING_LIB),
+            ),
+            SolidityAbi::V2 => (
+                String::from(CONTRACT_TEMPLATE_V2),
+                String::from(SOLIDITY_PAIRING_LIB_V2),
+            ),
+        };
+
+        let vk_regex = Regex::new(r#"(<%vk_[^i%]*%>)"#).unwrap();
+        let vk_gamma_abc_len_regex = Regex::new(r#"(<%vk_gamma_abc_length%>)"#).unwrap();
+        let vk_gamma_abc_repeat_regex = Regex::new(r#"(<%vk_gamma_abc_pts%>)"#).unwrap();
+        let vk_input_len_regex = Regex::new(r#"(<%vk_input_length%>)"#).unwrap();
+        let input_loop = Regex::new(r#"(<%input_loop%>)"#).unwrap();
+        let input_argument = Regex::new(r#"(<%input_argument%>)"#).unwrap();
+
+        template_text = vk_regex
+            .replace(template_text.as_str(), vk.alpha.to_string().as_str())
+            .into_owned();
+
+        template_text = vk_regex
+            .replace(template_text.as_str(), vk.beta.to_string().as_str())
+            .into_owned();
+
+        template_text = vk_regex
+            .replace(template_text.as_str(), vk.gamma.to_string().as_str())
+            .into_owned();
+
+        template_text = vk_regex
+            .replace(template_text.as_str(), vk.delta.to_string().as_str())
+            .into_owned();
+
+        let gamma_abc_count: usize = vk.gamma_abc.len();
+        template_text = vk_gamma_abc_len_regex
+            .replace(
+                template_text.as_str(),
+                format!("{}", gamma_abc_count).as_str(),
+            )
+            .into_owned();
+
+        template_text = vk_input_len_regex
+            .replace(
+                template_text.as_str(),
+                format!("{}", gamma_abc_count - 1).as_str(),
+            )
+            .into_owned();
+
+        // feed input values only if there are any
+        template_text = if gamma_abc_count > 1 {
+            input_loop.replace(
+                template_text.as_str(),
+                r#"
+        for(uint i = 0; i < input.length; i++){
+            inputValues[i] = input[i];
+        }"#,
+            )
+        } else {
+            input_loop.replace(template_text.as_str(), "")
+        }
+        .to_string();
+
+        // take input values as argument only if there are any
+        template_text = if gamma_abc_count > 1 {
+            input_argument.replace(
+                template_text.as_str(),
+                format!(", uint[{}] memory input", gamma_abc_count - 1).as_str(),
+            )
+        } else {
+            input_argument.replace(template_text.as_str(), "")
+        }
+        .to_string();
+
+        let mut gamma_abc_repeat_text = String::new();
+        for (i, g1) in vk.gamma_abc.iter().enumerate() {
+            gamma_abc_repeat_text.push_str(
+                format!(
+                    "vk.gamma_abc[{}] = Pairing.G1Point({});",
+                    i,
+                    g1.to_string().as_str()
+                )
+                .as_str(),
+            );
+            if i < gamma_abc_count - 1 {
+                gamma_abc_repeat_text.push_str("\n        ");
+            }
+        }
+
+        template_text = vk_gamma_abc_repeat_regex
+            .replace(template_text.as_str(), gamma_abc_repeat_text.as_str())
+            .into_owned();
+
+        let re = Regex::new(r"(?P<v>0[xX][0-9a-fA-F]{64})").unwrap();
+        template_text = re.replace_all(&template_text, "uint256($v)").to_string();
+
+        format!(
+            "{}{}{}",
+            SOLIDITY_G2_ADDITION_LIB, solidity_pairing_lib, template_text
+        )
+    }
+}
+
+const CONTRACT_TEMPLATE_V2: &str = r#"
+contract Verifier {
+    using Pairing for *;
+    struct VerifyingKey {
+        Pairing.G1Point alpha;
+        Pairing.G2Point beta;
+        Pairing.G2Point gamma;
+        Pairing.G2Point delta;
+        Pairing.G1Point[] gamma_abc;
+    }
+    struct Proof {
+        Pairing.G1Point a;
+        Pairing.G2Point b;
+        Pairing.G1Point c;
+    }
+    function verifyingKey() pure internal returns (VerifyingKey memory vk) {
+        vk.alpha = Pairing.G1Point(<%vk_alpha%>);
+        vk.beta = Pairing.G2Point(<%vk_beta%>);
+        vk.gamma = Pairing.G2Point(<%vk_gamma%>);
+        vk.delta = Pairing.G2Point(<%vk_delta%>);
+        vk.gamma_abc = new Pairing.G1Point[](<%vk_gamma_abc_length%>);
+        <%vk_gamma_abc_pts%>
+    }
+    function verify(uint[] memory input, Proof memory proof) internal view returns (uint) {
+        uint256 snark_scalar_field = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
+        VerifyingKey memory vk = verifyingKey();
+        require(input.length + 1 == vk.gamma_abc.length);
+        // Compute the linear combination vk_x
+        Pairing.G1Point memory vk_x = Pairing.G1Point(0, 0);
+        for (uint i = 0; i < input.length; i++) {
+            require(input[i] < snark_scalar_field);
+            vk_x = Pairing.addition(vk_x, Pairing.scalar_mul(vk.gamma_abc[i + 1], input[i]));
+        }
+        vk_x = Pairing.addition(vk_x, vk.gamma_abc[0]);
+        if(!Pairing.pairingProd4(
+             proof.a, proof.b,
+             Pairing.negate(vk_x), vk.gamma,
+             Pairing.negate(proof.c), vk.delta,
+             Pairing.negate(vk.alpha), vk.beta)) return 1;
+        return 0;
+    }
+    function verifyTx(
+            Proof memory proof<%input_argument%>
+        ) public view returns (bool r) {
+        uint[] memory inputValues = new uint[](input.length);
+        <%input_loop%>
+        if (verify(inputValues, proof) == 0) {
+            return true;
+        } else {
+            return false;
+        }
+    }
+}
+"#;
+
+const CONTRACT_TEMPLATE: &str = r#"
+contract Verifier {
+    using Pairing for *;
+    struct VerifyingKey {
+        Pairing.G1Point alpha;
+        Pairing.G2Point beta;
+        Pairing.G2Point gamma;
+        Pairing.G2Point delta;
+        Pairing.G1Point[] gamma_abc;
+    }
+    struct Proof {
+        Pairing.G1Point a;
+        Pairing.G2Point b;
+        Pairing.G1Point c;
+    }
+    function verifyingKey() pure internal returns (VerifyingKey memory vk) {
+        vk.alpha = Pairing.G1Point(<%vk_alpha%>);
+        vk.beta = Pairing.G2Point(<%vk_beta%>);
+        vk.gamma = Pairing.G2Point(<%vk_gamma%>);
+        vk.delta = Pairing.G2Point(<%vk_delta%>);
+        vk.gamma_abc = new Pairing.G1Point[](<%vk_gamma_abc_length%>);
+        <%vk_gamma_abc_pts%>
+    }
+    function verify(uint[] memory input, Proof memory proof) internal view returns (uint) {
+        uint256 snark_scalar_field = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
+        VerifyingKey memory vk = verifyingKey();
+        require(input.length + 1 == vk.gamma_abc.length);
+        // Compute the linear combination vk_x
+        Pairing.G1Point memory vk_x = Pairing.G1Point(0, 0);
+        for (uint i = 0; i < input.length; i++) {
+            require(input[i] < snark_scalar_field);
+            vk_x = Pairing.addition(vk_x, Pairing.scalar_mul(vk.gamma_abc[i + 1], input[i]));
+        }
+        vk_x = Pairing.addition(vk_x, vk.gamma_abc[0]);
+        if(!Pairing.pairingProd4(
+             proof.a, proof.b,
+             Pairing.negate(vk_x), vk.gamma,
+             Pairing.negate(proof.c), vk.delta,
+             Pairing.negate(vk.alpha), vk.beta)) return 1;
+        return 0;
+    }
+    function verifyTx(
+            uint[2] memory a,
+            uint[2][2] memory b,
+            uint[2] memory c<%input_argument%>
+        ) public view returns (bool r) {
+        Proof memory proof;
+        proof.a = Pairing.G1Point(a[0], a[1]);
+        proof.b = Pairing.G2Point([b[0][0], b[0][1]], [b[1][0], b[1][1]]);
+        proof.c = Pairing.G1Point(c[0], c[1]);
+        uint[] memory inputValues = new uint[](<%vk_input_length%>);
+        <%input_loop%>
+        if (verify(inputValues, proof) == 0) {
+            return true;
+        } else {
+            return false;
+        }
+    }
+}
+"#;
