+++ zokrates_core/src/proof_system/scheme/mod.rs
@@ -1,0 +1,17 @@
+use proof_system::solidity::SolidityAbi;
+use serde::de::DeserializeOwned;
+use serde::Serialize;
+use zokrates_field::Field;
+
+pub mod gm17;
+pub mod groth16;
+pub mod pghr13;
+
+pub trait Scheme<T: Field> {
+    type VerificationKey: Serialize + DeserializeOwned;
+    type ProofPoints: Serialize + DeserializeOwned;
+}
+
+pub trait SolidityCompatibleScheme<T: Field>: Scheme<T> {
+    fn export_solidity_verifier(vk: Self::VerificationKey, abi: SolidityAbi) -> String;
+}
